perf: Don't import Panel and Pandas in IPython
